requirements.txt
matplotlib
streamlit
google-generativeai
requests
matplotlib


